# packages/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: RNR API
  version: 0.1.0
servers:
  - url: http://localhost:3000
paths:
  /api/users:
    get:
      tags: [Users]
      summary: List all users
      responses:
        '200':
          description: Users list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
    post:
      tags: [Users]
      summary: Create new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserInput'
      responses:
        '200':
          description: Created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /api/users/{id}:
    get:
      tags: [Users]
      summary: Get user by ID
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: Not found
    put:
      tags: [Users]
      summary: Update user
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserInput'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: Not found
    delete:
      tags: [Users]
      summary: Delete user
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '404':
          description: Not found
  /api/posts:
    get:
      tags: [Posts]
      summary: List all posts
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: pageSize
          schema:
            type: integer
      responses:
        '200':
          description: Posts list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Post'
                  page:
                    type: integer
                  pageSize:
                    type: integer
                  total:
                    type: integer
    post:
      tags: [Posts]
      summary: Create new post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostInput'
      responses:
        '200':
          description: Created post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
  /api/posts/{id}:
    get:
      tags: [Posts]
      summary: Get post by ID
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '404':
          description: Not found
    put:
      tags: [Posts]
      summary: Update post
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostInput'
      responses:
        '200':
          description: Updated post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '404':
          description: Not found
    delete:
      tags: [Posts]
      summary: Delete post
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '404':
          description: Not found
  /api/posts/author/{authorId}:
    get:
      tags: [Posts]
      summary: Get posts by author
      parameters:
        - in: path
          name: authorId
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Posts by author
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Post'
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        name:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Post:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
          nullable: true
        published:
          type: boolean
        createdAt:
          type: string
          format: date-time
        authorId:
          type: integer
        author:
          $ref: '#/components/schemas/User'
    CreateUserInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
        name:
          type: string
    UpdateUserInput:
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
    CreatePostInput:
      type: object
      required: [title, authorId]
      properties:
        title:
          type: string
        content:
          type: string
        authorId:
          type: integer
        published:
          type: boolean
    UpdatePostInput:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        authorId:
          type: integer
        published:
          type: boolean
    UserResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/User'
    PostResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Post'