# packages/api-spec/src/gen/types/Post.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { User } from "./User.ts";

export type Post = {
    /**
     * @type integer | undefined
    */
    id?: number;
    /**
     * @type string | undefined
    */
    title?: string;
    /**
     * @type string
    */
    content?: string | null;
    /**
     * @type boolean | undefined
    */
    published?: boolean;
    /**
     * @type string | undefined, date-time
    */
    createdAt?: string;
    /**
     * @type integer | undefined
    */
    authorId?: number;
    /**
     * @type object | undefined
    */
    author?: User;
};