# packages/api-spec/src/gen/types/UpdatePostInput.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export type UpdatePostInput = {
    /**
     * @type string | undefined
    */
    title?: string;
    /**
     * @type string | undefined
    */
    content?: string;
    /**
     * @type integer | undefined
    */
    authorId?: number;
    /**
     * @type boolean | undefined
    */
    published?: boolean;
};